requirements.txt
pynvml
PyYAML==6.0.1
fire==0.5.0